Redo master-DB-stickiness in a different way.
Now, instead of a post-save hook, assume any POST request is a write, and set a cookie that keeps reads on the master for 15 seconds thereafter. Assumes we manually call using() to target reads to the master for the remainder of a request after a write. Also, renamed everything from "sticky" to "pinned" by popular demand.