Allows views to indicate that they do--or have done--a DB write.
Pay attention to `response._db_write` and provide a decorator to set it to
True. If the attribute exists and is true, set the pinning cookie as if the
request had been a `POST`.

If a view _may_ do a DB write, set the attribute on the response object
if necessary before returning it.